dpa_offload: Resync dpa_classifier w/ SDK 1.5
Updated the dpa_classifier component to bring it to the version
delivered with SDK 1.5.

Reviewed-on: http://git.am.freescale.net:8181/8716
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>

+++ drivers/staging/fsl_dpa_offload/dpa_classifier_ioctl.h
@@ -364,7 +364,7 @@ struct compat_ioc_dpa_cls_hm_vlan_params {
 struct compat_ipv4_header {
 	struct iphdr			header;
 	compat_uptr_t			options;
-	unsigned int			options_size;
+	uint8_t				options_size;
 };
 
 struct dpa_cls_compat_hm_nat_pt_params {
